backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.3.2
google-cloud-vision>=3.4.5
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
pydantic>=2.5.0
reportlab>=4.0.0
gtts>=2.3.0
PyPDF2>=3.0.0
